Switch to named graph queries
The queries without graph specified are very ineffective.
On a phone getting a list of 300+ pictures took almost half a minute.
With named graph it's semi-immediate.

As side-effect of the separate graphs, not all data is anymore available
for or root types. If app developer asks for files, it's not possible
to get anymore duration, artist or such metadata properties.

On the implementation side, there's now some separation between
FileSystem graph and other graph properties. On the FileSystem the
main items are file urls while elsewhere they are urn that have
reference to a file.

Some other small changes:
- "path" property removed. There's already filePath which can be used
for path needs. The belongsTo type of property on the implemenation is
not necessarily available in all the graphs and where it is it's used
for other type of containers too.
- fileLastAccessed removed. At the moment available only in FileSystem
graph, and the tracking of access time is often disabled or throttled
on the mount level. fileLastModified should be often good enough
replacement.